Use an editor model for the event editor.
Anything that is editable now has its own item in a model that is on the
scope.  This is what is bound in the view.  All manipulation and validation
is performed against this model.  At the time of save, then, the data that
is in the editor model is copied back to the data model in the proper format
and type.