Save the category as a string.
The category is saved as a string, even if it was selected from one of the objects in the list.
The typeahead list is really just treated as a list of values used before, but there is no real
relation between the event categories at the category for the event (at least not at this time).