Use independent fields on the model.
With all of the manipulation that can be done out of the watchers, the end
date was getting out of sync with the value used in the picker.  A watcher
needed to be added to the directive, but this only works well if the
ngModel used for the picker is only used on a single picker.  Otherwise,
things can get confusing.